+++ Libft/ft_putendl_fd.c
@@ -6,7 +6,7 @@
 /*   By: jaehpark <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/12 21:15:42 by jaehpark          #+#    #+#             */
-/*   Updated: 2020/11/22 02:04:06 by jaehpark         ###   ########.fr       */
+/*   Updated: 2020/11/22 19:26:36 by jaehpark         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putendl_fd(char *s, int fd)
 {
-	size_t	i;
+	int	i;
 
 	if (!s)
 		return ;
