+++ Libft/ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: jaehpark <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/12 21:15:05 by jaehpark          #+#    #+#             */
-/*   Updated: 2020/11/22 02:03:55 by jaehpark         ###   ########.fr       */
+/*   Updated: 2020/11/22 19:26:52 by jaehpark         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(char *s, int fd)
 {
-	size_t	i;
+	int	i;
 
 	if (!s)
 		return ;
